fix: ensure CA cert generation respects the hour flag
This PR fixes a bug with cert generation via `osctl gen ca`. The value
specified by the --hour flag was never being appended to the CA options
and also, since the default value for `hour` was being set on init, the
CA default was being overwritten by a subsequent command setup for the
CRT default (24h). This PR moves to using two distinct variables for
those hour values. Will fix #1911.

Signed-off-by: Spencer Smith <robertspencersmith@gmail.com>